+++ 06week/ticTacToe/index.html
@@ -23,7 +23,6 @@
   </div>
   <div id="announce-winner"></div>
   <button id="clear">Clear Board</button>
-  <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-  <script type="text/javascript" src="/script2.js"></script>
+  <script type="text/javascript" src="./script.js" />
 </body>
 </html>
