+++ 06week/towersOfHanoi/index.html
@@ -17,7 +17,7 @@
   <div data-stack="3">
   </div>
   <div id="announce-game-won"></div>
-  <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-  <script type="text/javascript" src="script.js"></script>
+
+  <script type="text/javascript" src="./script.js"/>
 </body>
 </html>
