+++ 06week/towersOfHanoi/script.js
@@ -1,29 +1,7 @@
 'use strict';
 
 document.addEventListener('DOMContentLoaded', () => {
-  // Your code here
 
-  const assert = require('assert');
-  const readline = require('readline');
-  const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout
-  });
-
-  let stacks = {
-    a: [4, 3, 2, 1],
-    b: [],
-    c: []
-  };
-
-  function printStacks() {
-    console.log("a: " + stacks.a);
-    console.log("b: " + stacks.b);
-    console.log("c: " + stacks.c);
-  }
-
-  //naming of origin column and destination column for each piece.
-  //this allows each block to be removed from the "end" of a column and placed into a different column
   function movePiece(startStack, endStack) {
     // Your code here
     let block = stacks[startStack].pop();
@@ -56,59 +34,6 @@ document.addEventListener('DOMContentLoaded', () => {
       movePiece(startStack, endStack);
     }
     checkForWin();
-  }
-
-  function getPrompt() {
-    printStacks();
-    rl.question('start stack: ', (startStack) => {
-      rl.question('end stack: ', (endStack) => {
-        towersOfHanoi(startStack, endStack);
-        getPrompt();
-      });
-    });
-  }
-
-// Tests -----------------------------------
-
-  if (typeof describe === 'function') {
-
-    describe('#towersOfHanoi()', () => {
-      it('should be able to move a block', () => {
-        towersOfHanoi('a', 'b');
-        assert.deepEqual(stacks, { a: [4, 3, 2], b: [1], c: [] });
-      });
-    });
-
-    describe('#isLegal()', () => {
-      it('should not allow an illegal move', () => {
-        stacks = {
-          a: [4, 3, 2],
-          b: [1],
-          c: []
-        };
-        assert.equal(isLegal('a', 'b'), false);
-      });
-      it('should allow a legal move', () => {
-        stacks = {
-          a: [4, 3, 2, 1],
-          b: [],
-          c: []
-        };
-        assert.equal(isLegal('a', 'c'), true);
-      });
-    });
-    describe('#checkForWin()', () => {
-      it('should detect a win', () => {
-        stacks = { a: [], b: [4, 3, 2, 1], c: [] };
-        assert.equal(checkForWin(), true);
-        stacks = { a: [1], b: [4, 3, 2], c: [] };
-        assert.equal(checkForWin(), false);
-      });
-    });
-  } else {
-
-    getPrompt();
-
   }
 
 });
